extraction of the links doesnt work but scraping does figure out how to stop drivers from opening because old solution no longer works
