+++ fashioncrawler/utils/__init__.py
@@ -1,0 +1,2 @@
+from .io_utils import IOUtils
+from .logger_config import configure_logger
