+++ fashioncrawler/utils/io_utils.py
